Reverting back to paper submodule.
Quality of life reversion back to using the hugo paper theme as a
submodule.

Hugo lets you override theme files by rooting the file with the same
name path in the theme, but just rooted at the root of your repo. This
is the intended way to modify a theme.

To revert back to this, I did the following:

1. get all the changes we made to themes/paper:
- diff current files in themes/paper with the source paper code
   - diffed with original repo corresponding to the recent commit in our
   repo for themes/paper NOT from smchase

2. delete themes/paper and all commit history
- used git-filter-repo as mentionned here: https://stackoverflow.com/a/61544937

3. bring paper back in as a submodule (point it to the same commit as we began with, `456c165`)

4. add the changed files from part 1 as modifications in root